tests/requirements.txt
pytest
pytest-cov
pytest-django
